requirements.txt
jupyter
pandas
numpy
sqlalchemy
scikit-learn
ftfy
psycopg2
python-Levenshtein
fuzzywuzzy
python-dotenv